refactor(sdk): Make ConnectionString protocol-agnostic with generic type
Previously, the implementation of ConnectionString and ConnectionStringOptions are tightly
coupled with TCP support, making it difficult to expand the support for other protocols, e.g.
QUIC, HTTP, etc. This created unwanted dependencies on protocol-specific details.
Note that only fields that are common for different protocols should have a getter/setter
method under ConnectionStringOptions.

To bring better flexibility and maintainbility, the following changes are made
	- Convert ConnectionStringOptions from struct to trait
	- Leverage protocol-specific implementation to configuration of each protocol
	- Make ConnectionString focus on string manipulation without too much concern of
	  protocol-specific knowledge

With this refactoring, ConnectionString now serves as a foundation that different protocols
can build upon by implementing the ConnectionStringOptions trait.
The TCP implementation has been ported to this new architecture of ConnectioinString.
Different protocols will be built on top of it in the coming patches.
