+++ core/common/src/lib.rs
@@ -59,6 +59,7 @@ pub use types::configuration::quick_config::quic_client_reconnection_config::*;
 pub use types::configuration::tcp_config::tcp_client_config::*;
 pub use types::configuration::tcp_config::tcp_client_config_builder::*;
 pub use types::configuration::tcp_config::tcp_client_reconnection_config::*;
+pub use types::configuration::tcp_config::tcp_connection_string_options::*;
 pub use types::confirmation::*;
 pub use types::consumer::consumer_group::*;
 pub use types::consumer::consumer_kind::*;
