+++ core/common/src/types/configuration/auth_config/connection_string.rs
@@ -16,22 +16,19 @@
  * under the License.
  */
 
-use crate::{
-    AutoLogin, ConnectionStringOptions, Credentials, IggyDuration, IggyError,
-    TcpClientReconnectionConfig,
-};
+use crate::{AutoLogin, ConnectionStringOptions, Credentials, IggyError};
 use std::str::FromStr;
 
 const CONNECTION_STRING_PREFIX: &str = "iggy://";
 
 #[derive(Debug)]
-pub struct ConnectionString {
+pub struct ConnectionString<T: ConnectionStringOptions + Default> {
     server_address: String,
     auto_login: AutoLogin,
-    options: ConnectionStringOptions,
+    options: T,
 }
 
-impl ConnectionString {
+impl<T: ConnectionStringOptions + Default> ConnectionString<T> {
     pub fn server_address(&self) -> &str {
         &self.server_address
     }
@@ -40,12 +37,10 @@ impl ConnectionString {
         &self.auto_login
     }
 
-    pub fn options(&self) -> &ConnectionStringOptions {
+    pub fn options(&self) -> &T {
         &self.options
     }
-}
 
-impl ConnectionString {
     pub fn new(connection_string: &str) -> Result<Self, IggyError> {
         if connection_string.is_empty() {
             return Err(IggyError::InvalidConnectionString);
@@ -83,7 +78,7 @@ impl ConnectionString {
             return Err(IggyError::InvalidConnectionString);
         }
 
-        if !server_address.contains(':') {
+        if !server_address.contains(':') || server_address.starts_with(':') {
             return Err(IggyError::InvalidConnectionString);
         }
 
@@ -98,9 +93,9 @@ impl ConnectionString {
 
         let connection_string_options;
         if let Some(options) = server_and_options.get(1) {
-            connection_string_options = ConnectionString::parse_options(options)?;
+            connection_string_options = T::parse_options(options)?;
         } else {
-            connection_string_options = ConnectionStringOptions::default();
+            connection_string_options = T::default();
         }
 
         Ok(ConnectionString {
@@ -112,89 +107,11 @@ impl ConnectionString {
             options: connection_string_options,
         })
     }
-
-    fn parse_options(options: &str) -> Result<ConnectionStringOptions, IggyError> {
-        let options = options.split('&').collect::<Vec<&str>>();
-        let mut tls_enabled = false;
-        let mut tls_domain = "localhost".to_string();
-        let mut tls_ca_file = None;
-        let mut reconnection_retries = "unlimited".to_owned();
-        let mut reconnection_interval = "1s".to_owned();
-        let mut reestablish_after = "5s".to_owned();
-        let mut heartbeat_interval = "5s".to_owned();
-        let mut nodelay = false;
-
-        for option in options {
-            let option_parts = option.split('=').collect::<Vec<&str>>();
-            if option_parts.len() != 2 {
-                return Err(IggyError::InvalidConnectionString);
-            }
-            match option_parts[0] {
-                "tls" => {
-                    tls_enabled = option_parts[1] == "true";
-                }
-                "tls_domain" => {
-                    tls_domain = option_parts[1].to_string();
-                }
-                "tls_ca_file" => {
-                    tls_ca_file = Some(option_parts[1].to_string());
-                }
-                "reconnection_retries" => {
-                    reconnection_retries = option_parts[1].to_string();
-                }
-                "reconnection_interval" => {
-                    reconnection_interval = option_parts[1].to_string();
-                }
-                "reestablish_after" => {
-                    reestablish_after = option_parts[1].to_string();
-                }
-                "heartbeat_interval" => {
-                    heartbeat_interval = option_parts[1].to_string();
-                }
-                "nodelay" => {
-                    nodelay = option_parts[1] == "true";
-                }
-                _ => {
-                    return Err(IggyError::InvalidConnectionString);
-                }
-            }
-        }
-
-        let reconnection = TcpClientReconnectionConfig {
-            enabled: true,
-            max_retries: match reconnection_retries.as_str() {
-                "unlimited" => None,
-                _ => Some(
-                    reconnection_retries
-                        .parse()
-                        .map_err(|_| IggyError::InvalidNumberValue)?,
-                ),
-            },
-            interval: IggyDuration::from_str(reconnection_interval.as_str())
-                .map_err(|_| IggyError::InvalidConnectionString)?,
-            reestablish_after: IggyDuration::from_str(reestablish_after.as_str())
-                .map_err(|_| IggyError::InvalidConnectionString)?,
-        };
-
-        let heartbeat_interval = IggyDuration::from_str(heartbeat_interval.as_str())
-            .map_err(|_| IggyError::InvalidConnectionString)?;
-
-        let connection_string_options = ConnectionStringOptions::new(
-            tls_enabled,
-            tls_domain,
-            tls_ca_file,
-            reconnection,
-            heartbeat_interval,
-            nodelay,
-        );
-
-        Ok(connection_string_options)
-    }
 }
 
-impl FromStr for ConnectionString {
+impl<T: ConnectionStringOptions + Default> FromStr for ConnectionString<T> {
     type Err = IggyError;
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        ConnectionString::new(s)
+        ConnectionString::<T>::new(s)
     }
 }
