+++ core/common/src/types/configuration/tcp_config/mod.rs
@@ -18,3 +18,4 @@
 pub(crate) mod tcp_client_config;
 pub(crate) mod tcp_client_config_builder;
 pub(crate) mod tcp_client_reconnection_config;
+pub(crate) mod tcp_connection_string_options;
